Update Main.java
Type long changed to int

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        long price = 1000;
+        int price = 1000;
         int coefficient = 20;
-        long bonus = price / coefficient;
+        int bonus = price / coefficient;
         System.out.println(bonus);
     }
 }
